+++ index.html
@@ -4,12 +4,12 @@
 <head>
   <meta charset="UTF-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>A Richmond's Minimal Surface | Andrei Chervoniak</title>
+  <title>A Sievert's Surface | Andrei Chervoniak</title>
   <link rel="stylesheet" href="./style.css" />
 </head>
 
 <body>
-  <h1>A Richmond's Minimal Surface</h1>
+  <h1>A Sievert's Surface Surface</h1>
   <p>
     Drag your mouse on the canvas to rotate it. (On a touch screen, you can
     use your finger.)
