+++ main.js
@@ -1,75 +1,15 @@
-import { SurfaceModel } from "./geometry.js";
+import { SurfaceModel } from "./SurfaceModel.js";
+import { ShaderProgram } from "./ShaderProgram.js";
 
 let gl;
 let surface;
 let program;
 let ball;
 
-function deg2rad(angle) {
-	return (angle * Math.PI) / 180;
-}
-
-class ShaderProgram {
-	constructor(name) {
-		this.name = name;
-	}
-
-	use(gl) {
-		gl.useProgram(this.prog);
-		this.setupAttributes(gl);
-		this.setupUniforms(gl);
-	}
-
-	setupAttributes(gl) {
-		this.vertexAttrib = gl.getAttribLocation(this.prog, "vertex");
-		this.normalAttrib = gl.getAttribLocation(this.prog, "normal");
-	}
-
-	setupUniforms(gl) {
-		this.matrixUni = gl.getUniformLocation(this.prog, "matrix");
-		this.normalMatrixUni = gl.getUniformLocation(this.prog, "normalMatrix");
-		this.lightDirectionUni = gl.getUniformLocation(this.prog, "lightDirection");
-		this.viewPositionUni = gl.getUniformLocation(this.prog, "viewPosition");
-		this.ambientColorUni = gl.getUniformLocation(this.prog, "ambientColor");
-		this.diffuseColorUni = gl.getUniformLocation(this.prog, "diffuseColor");
-		this.specularColorUni = gl.getUniformLocation(this.prog, "specularColor");
-		this.shininessUni = gl.getUniformLocation(this.prog, "shininess");
-	}
-
-	init(gl, vs, fs) {
-		const vsh = this.compileShader(gl, gl.VERTEX_SHADER, vs);
-		const fsh = this.compileShader(gl, gl.FRAGMENT_SHADER, fs);
-		this.prog = this.createProgram(gl, vsh, fsh);
-	}
-
-	compileShader(gl, type, source) {
-		const shader = gl.createShader(type);
-		gl.shaderSource(shader, source);
-		gl.compileShader(shader);
-		if (!gl.getShaderParameter(shader, gl.COMPILE_STATUS)) {
-			throw new Error(
-				`Shader compilation error: ${gl.getShaderInfoLog(shader)}`,
-			);
-		}
-		return shader;
-	}
-
-	createProgram(gl, vsh, fsh) {
-		const prog = gl.createProgram();
-		gl.attachShader(prog, vsh);
-		gl.attachShader(prog, fsh);
-		gl.linkProgram(prog);
-		if (!gl.getProgramParameter(prog, gl.LINK_STATUS)) {
-			throw new Error(`Program link error: ${gl.getProgramInfoLog(prog)}`);
-		}
-		return prog;
-	}
-}
-
 function getUVSteps() {
 	return {
-		u: parseInt(document.getElementById("u-stepper").value, 10),
-		v: parseInt(document.getElementById("v-stepper").value, 10),
+		u: Number.parseInt(document.getElementById("u-stepper").value, 10),
+		v: Number.parseInt(document.getElementById("v-stepper").value, 10),
 	};
 }
 
@@ -129,7 +69,7 @@ function draw() {
 	gl.clearColor(1, 1, 1, 1);
 	gl.clear(gl.COLOR_BUFFER_BIT | gl.DEPTH_BUFFER_BIT);
 
-	const projection = m4.perspective(Math.PI / 8, 1, 8, 12);
+	const projection = m4.perspective(Math.PI / 8, 1, 0.1, 100);
 	const modelView = ball.getViewMatrix();
 
 	const rotateToPointZero = m4.axisRotation(
@@ -138,13 +78,13 @@ function draw() {
 	);
 	const translateToPointZero = m4.translation(0, 0, -10);
 
-	const matAccum0 = m4.multiply(rotateToPointZero, modelView);
-	const matAccum1 = m4.multiply(translateToPointZero, matAccum0);
+	const matAcc0 = m4.multiply(rotateToPointZero, modelView);
+	const matAcc1 = m4.multiply(translateToPointZero, matAcc0);
 
-	const modelViewProjection = m4.multiply(projection, matAccum1);
+	const modelViewProjection = m4.multiply(projection, matAcc1);
 	gl.uniformMatrix4fv(program.matrixUni, false, modelViewProjection);
 
-	const normalMatrix = m4.transpose(m4.inverse(matAccum1));
+	const normalMatrix = m4.transpose(m4.inverse(matAcc1));
 	gl.uniformMatrix4fv(program.normalMatrixUni, false, normalMatrix);
 
 	gl.uniform3fv(program.viewPositionUni, [0.0, 0.0, 5.0]);
